PowerUp and PowerDown containers, routes for powerup and powerdown
